graph/pyproject.toml
[project]
name = "graph"
version = "0.1.0"
description = "LangGraph Agent for BookedAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]
